+++ markdown.d.ts
@@ -970,6 +970,10 @@ export namespace MarkdownParser {
       Inline<'mark'>,
       Parser<HTMLElement | string, Context, [
         InlineParser,
+        Parser<HTMLElement | string, Context, [
+          InlineParser,
+          InlineParser,
+        ]>,
       ]> {
     }
     export interface EmStrongParser extends
@@ -977,15 +981,21 @@ export namespace MarkdownParser {
       Inline<'emstrong'>,
       Parser<HTMLElement | string, Context, [
         InlineParser,
+        Parser<HTMLElement | string, Context, [
+          InlineParser,
+          InlineParser,
+        ]>,
       ]> {
     }
     export interface StrongParser extends
       // **abc**
       Inline<'strong'>,
       Parser<HTMLElement | string, Context, [
-        EmphasisParser,
         InlineParser,
-        SourceParser.StrParser,
+        Parser<HTMLElement | string, Context, [
+          InlineParser,
+          InlineParser,
+        ]>,
       ]> {
     }
     export interface EmphasisParser extends
@@ -994,6 +1004,10 @@ export namespace MarkdownParser {
       Parser<HTMLElement | string, Context, [
         StrongParser,
         InlineParser,
+        Parser<HTMLElement | string, Context, [
+          InlineParser,
+          InlineParser,
+        ]>,
       ]> {
     }
     export interface CodeParser extends