+++ src/parser/inline/strong.ts
@@ -1,7 +1,6 @@
 import { StrongParser } from '../inline';
-import { union, some, creator, surround, close, lazy } from '../../combinator';
+import { union, sequence, some, creator, surround, close, lazy } from '../../combinator';
 import { inline } from '../inline';
-import { emphasis } from './emphasis';
 import { str } from '../source';
 import { startTight, isEndTightNodes } from '../util';
 import { html, defrag } from 'typed-dom';
@@ -9,7 +8,10 @@ import { unshift } from 'spica/array';
 
 export const strong: StrongParser = lazy(() => creator(surround(close(
   str('**'), /^(?!\*)/),
-  startTight(some(union([emphasis, some(inline, '*'), str('*')]), '**')),
+  startTight(some(union([
+    some(inline, /^\s*\*\*/),
+    sequence([some(inline, '*'), inline]),
+  ]))),
   str('**'), false,
   ([as, bs, cs], rest) =>
     isEndTightNodes(bs)